Move sideload to driver Move formData to utils Move screenshot functions to utils
